requirements.txt
scipy~=1.2.0
numpy~=1.18.4
tensorflow~=1.15.2
matplotlib~=3.1.1
pandas~=0.25.2
ipython~=7.19.0
Pillow~=8.0.1
torchvision~=0.8.1+cu110
torch~=1.7.0+cu110
opencv-python~=4.2.0.34
scikit-learn~=0.23.2
